# perf: `fn rav1d_cdef_brow`: reduce stack usage