@ohadravid 2025-06-10
@kkysen -

1. ✅ moved the `backup2x8` to a separate PR 
2. ❓ w.r.t only changing `bot` vs all 3 variables: in my tests, only changing all of them gives the perf boost - https://github.com/memorysafety/rav1d/pull/1418#discussion_r2136944989