@ohadravid 2025-06-13
@kkysen - added the utlity functions - let me know if you like them (can update the names / docs / etc).

Also tried to apply the same trick everywhere https://github.com/memorysafety/rav1d/pull/1418#discussion_r2145087617 .