@ohadravid 2025-06-14
@kkysen do you want me to split [680eb25](https://github.com/memorysafety/rav1d/pull/1418/commits/680eb250c3ba5cf5e6e983730a97532d0323acf5) and the followup update to the other callsites to a new PR?